@@ -1,16 +1,8 @@
 {
   "name": "js-mdict",
-  "version": "5.0.5",
+  "version": "6.0.0-alpha-1",
   "description": "mdict (*.mdx, *.mdd) file reader",
-  "main": "lib/mdict.js",
-  "types": "typings/mdict.d.ts",
-  "scripts": {
-    "bench": "npm run build && mocha --require @babel/register test/decode_key_block.bench.js",
-    "test": "npm run build && jest test/*.spec.js",
-    "build": "babel src --out-dir lib",
-    "lint": "eslint src",
-    "format": "prettier -w src/*.js"
-  },
+  "author": "terasum <terasum@163.com>",
   "repository": "git+https://github.com/terasum/js-mdict.git",
   "keywords": [
     "mdict",
@@ -20,38 +12,88 @@
     "vocabulary",
     "borwser"
   ],
-  "author": "terasum <terasum@163.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/terasum/js-mdict/issues"
   },
   "homepage": "https://github.com/terasum/js-mdict#readme",
-  "dependencies": {
-    "@babel/runtime": "^7.25.6",
-    "@xmldom/xmldom": "^0.8.6",
-    "bl": "^2.2.1",
-    "buffer-to-arraybuffer": "0.0.6",
-    "lemmatizer": "0.0.1",
-    "levenshtein-edit-distance": "^3.0.1",
-    "read-chunk": "^3.2.0"
+  "type": "module",
+  "exports": {
+    "require": "./dist/cjs/index.js",
+    "import":  "./dist/esm/index.js"
   },
+  "main": "./dist/cjs/index.js",
+  "types": "./dist/cjs/index.d.ts",
   "engines": {
-    "node": ">=6.0.0"
+    "node": ">=10.20.0"
+  },
+  "scripts": {
+    "prepareBuild": "shx mkdir -p dist/build",
+    "prepublishOnly": "npm run prepareBuild && npm run build",
+    "clean": "shx rm -rf dist/*",
+    "test": "jest",
+    "build-cjs": "tsc --module commonjs --outDir dist/cjs/ && echo '{\"type\": \"commonjs\"}' > dist/cjs/package.json",
+    "build-esm": "tsc  --module nodenext --moduleResolution nodenext --outDir dist/esm/ && echo '{\"type\": \"module\"}' > dist/cjs/package.json",
+    "build": "npm run build-cjs && npm run build-esm",
+    "watch": "tsc --watch",
+    "testall": "mocha --ui qunit dist/test",
+    "lint": "eslint --quiet 'src/**/*.ts' 'test/**/*.ts'",
+    "lint-fix": "eslint --quiet 'src/**/*.ts' 'test/**/*.ts' --fix",
+    "prepack": "npm run clean && npm run build",
+    "bench": "npm run build && mocha --require @babel/register test/decode_key_block.bench.js"
+  },
+  "dependencies": {
+    "@types/pako": "^2.0.3",
+    "@xmldom/xmldom": "^0.8.6",
+    "levenshtein-edit-distance": "^3.0.1",
+    "lodash": "^4.17.21",
+    "lzox1-rs": "link:src/lzox1-rs/pkg",
+    "pako": "^2.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.21.0",
-    "@babel/core": "^7.21.3",
-    "@babel/plugin-transform-arrow-functions": "^7.20.7",
-    "@babel/plugin-transform-runtime": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/register": "^7.21.0",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.12.0",
+    "@jest/globals": "^29.7.0",
+    "@types/chai": "^5.0.0",
+    "@types/jest": "^29.5.13",
+    "@types/lodash": "^4.17.10",
+    "@types/node": "^22.7.3",
+    "@types/sinon": "^17.0.3",
+    "@typescript-eslint/eslint-plugin": "^8.8.0",
+    "@typescript-eslint/parser": "^8.8.0",
     "benchmark": "^2.1.4",
-    "chai": "^4.2.0",
-    "eslint": "^8.36.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.5.0",
-    "mocha": "^10.2.0",
-    "prettier": "^2.8.7"
+    "chai": "^5.1.1",
+    "eslint": "^9.12.0",
+    "globals": "^15.10.0",
+    "jest": "^29.7.0",
+    "nyc": "^17.1.0",
+    "shx": "^0.3.4",
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.6.0"
+  },
+  "files": [
+    "package.json",
+    "README.md",
+    "dist",
+    "LICENSE",
+    "src"
+  ],
+  "mocha": {
+    "loader": [
+      "ts-node/esm"
+    ],
+    "recursive": true,
+    "extension": [
+      "ts",
+      "js",
+      "tsx"
+    ],
+    "exit": true
+  },
+  "nyc": {
+    "extension": [
+      ".ts"
+    ]
   }
 }
