@@ -1,19 +1,9 @@
-/// <reference path="../typings/MdictBase.d.ts">
-
-import readChunk from 'read-chunk';
 import assert from 'assert';
-import BufferList from 'bl';
-// use nodejs embbed zlib instead of pako, only under nodejs
-// use pako = 347ms, use zlib = 290ms
-import zlib from 'zlib';
-const pako = {};
-pako.inflate = zlib.inflateSync;
-
-import bufferToArrayBuffer from 'buffer-to-arraybuffer';
 import common from './utils';
-import lzo1x from './lzo-wrapper';
-
-import measure from './measure_util';
+import lzo1x from './lzo1x-wrapper';
+import { NumFmt, readChunkSync } from './utils';
+import pako from "pako";
+import zlib from "zlib";
 
 const UTF_16LE_DECODER = new TextDecoder('utf-16le');
 const UTF16 = 'UTF-16';
@@ -27,21 +17,180 @@ const BIG5 = 'BIG5';
 const GB18030_DECODER = new TextDecoder('gb18030');
 const GB18030 = 'GB18030';
 
-const BASE64ENCODER = function (arrayBuffer) {
-  return arrayBuffer.toString('base64');
+const BASE64ENCODER = function (arrayBuffer: ArrayBuffer): string {
+  return Buffer.from(arrayBuffer).toString('base64');
 };
 
+export interface MDictOptions {
+  passcode?: string;
+  debug?: boolean;
+  resort?: boolean;
+  isStripKey?: boolean;
+  isCaseSensitive?: boolean;
+}
+
+export interface MDictHeader {
+  [key: string]: string;
+}
+
+export interface KeyHeader {
+  keyBlocksNum: number;
+  entriesNum: number;
+  keyBlockInfoDecompSize: number;
+  keyBlockInfoCompSize: number;
+  keyBlocksTotalSize: number;
+}
+
+export interface KeyBlockInfo {
+  firstKey: string;
+  lastKey: string;
+  keyBlockCompSize: number;
+  keyBlockCompAccumulator: number;
+  keyBlockDecompSize: number;
+  keyBlockDecompAccumulator: number;
+  keyBlockEntriesNum: number;
+  keyBlockEntriesAccumulator: number;
+  keyBlockIndex: number;
+}
+
+export interface RecordHeader {
+  recordBlocksNum: number;
+  entriesNum: number;
+  recordBlockInfoCompSize: number;
+  recordBlockCompSize: number;
+}
+
+export interface RecordBlockInfo {
+  compSize: number;
+  compAccumulator: number;
+  decompSize: number;
+  decompAccumulator: number;
+}
+
+export interface KeyListItem {
+  recordStartOffset: number;
+  keyText: string;
+  keyBlockIdx: number;
+  original_idx?: number;
+  nextRecordStartOffset?: number;
+}
+
+export interface KeyRecord {
+  keyText: string;
+  recordStartOffset: number;
+  recordOffset?: number;
+  rofset?: number;
+  nextRecordStartOffset?: number;
+  original_idx?: number;
+}
+
 /**
+ * @class MdictBase, the basic mdict diction parser class
+ * @brif
+ * STEPS:
+ * 1. read mdict file header
+ * 2. read key header
+ * 3. read key block info
+ * 4. read key block
+ * 5. read record header
+ * 6. read record block
  *
- * class MdictBase, the basic mdict diction parser class
+ * 词典结构包括如下部分:
+ *
+ * Header     : 记录词典的meta信息，包括名称、描述、样式、编码方式等
+ * KeyInfo    : 记录词典的Key排列信息，设计用于索引
+ * KeyBlock   : 记录词典的所有key列表信息，可以在 key block 中得到本词典的所有词条
+ * RecordInfo : 记录词典的所有record词条释义信息，可以加速检索
+ * RecordBlock: 记录词典的所有record词条释义，如果是mdd文件，则为二进制图片、音频等
+ *
+ * @return {MDict}
  */
 class MDictBase {
+  // mdx/mdd 文件名
+  fname: string;
+  // mdx 密码
+  _passcode: string | undefined;
+  // 当前文件读取偏移指针
+  _offset: number;
+  // ext 文件后缀
+  ext: string;
+  // options 读取选项
+  options: MDictOptions;
+  // header 开始偏移
+  _headerStartOffset: number;
+  // header 结束偏移
+  _headerEndOffset: number;
+  // header 数据
+  header: MDictHeader;
+
+  // keyHeader 开始偏移
+  _keyHeaderStartOffset: number;
+  // keyHeader 结束偏移
+  _keyHeaderEndOffset: number;
+  // keyHeader 数据
+  keyHeader: KeyHeader;
+
+  // keyBlockInfo 开始偏移
+  _keyBlockInfoStartOffset: number;
+  // keyBlockInfo 结束偏移
+  _keyBlockInfoEndOffset: number;
+  // keyBlockInfo 数据 (Key Block Info list)
+  keyBlockInfoList: KeyBlockInfo[];
+
+  // keyBlock 开始偏移
+  _keyBlockStartOffset: number;
+  // keyBlock 结束偏移
+  _keyBlockEndOffset: number;
+  // keyList 数据(词条列表)
+  keyList: KeyListItem[];
+
+  // recordHeader 开始偏移
+  _recordHeaderStartOffset: number;
+  // recordHeader 结束偏移
+  _recordHeaderEndOffset: number;
+  // recordHeader 数据
+  recordHeader: RecordHeader;
+
+  // recordInfo 开始偏移
+  _recordBlockInfoStartOffset: number;
+  // recordInfo 结束偏移
+  _recordBlockInfoEndOffset: number;
+  // recordBlockInfo 数据
+  recordBlockInfoList: RecordBlockInfo[];
+
+  // recordBlock 开始偏移
+  _recordBlockStartOffset: number;
+  // recordBlock 结束偏移
+  _recordBlockEndOffset: number;
+  // keyData 数据
+  recordBlockDataList: any[];
+
+  // mdx version
+  _version: number = 0;
+  // num width
+  _numWidth: number = 0;
+  // num format
+  _numFmt: NumFmt = common.NUMFMT_UINT32 as NumFmt;
+  // encoding 编码
+  _encoding: string = '';
+  // decoder 解码器
+  _decoder: TextDecoder = new TextDecoder();
+  // 是否加密
+  _encrypt: number = 0;
+
+  // keyListRemap 重定位map
+  keyListRemap: { [key: number]: number } = {};
+
   /**
    * mdict constructor
    * @param {string} fname
    * @param {string} passcode
    */
-  constructor(fname, passcode, options) {
+  constructor(
+    fname: string,
+    passcode?: string,
+    options?: Partial<MDictOptions>
+  ) {
     // the mdict file name
     this.fname = fname;
     // the dictionary file decrypt pass code
@@ -69,6 +218,10 @@ class MDictBase {
     this.header = {};
     this._readHeader();
 
+    if (this.options.debug) {
+      console.log('readHeader finished', { keyHeader: this.header });
+    }
+
     // -------------------------
     // dict key header section
     // --------------------------
@@ -84,6 +237,10 @@ class MDictBase {
     // read key header
     this._readKeyHeader();
 
+    if (this.options.debug) {
+      console.log('readKeyHeader finished', { keyHeader: this.keyHeader });
+    }
+
     // -------------------------
     // dict key info section
     // --------------------------
@@ -93,67 +250,90 @@ class MDictBase {
     this.keyBlockInfoList = [];
     this._readKeyBlockInfo();
 
+    if (this.options.debug) {
+      const tempKeyBlockInfoList = this.keyBlockInfoList.slice(0, 10);
+      console.log('readKeyBlockInfo finished', {
+        keyBlockInfoList_0_10: tempKeyBlockInfoList,
+      });
+    }
+
     // -------------------------
     // dict key block section
     // --------------------------
     this._keyBlockStartOffset = 0;
     this._keyBlockEndOffset = 0;
+    this.keyList = [];
+
+    // @depreciated
+    // decodeKeyBlock method is very slow, avoid invoke dirctly
+    // this method will return the whole words list of the dictionaries file, this is very slow
+    // operation, and you should do this background, or concurrently.
+    // NOTE: this method is wrapped by method medict.RangeWords();
+    // NOTE: 本方法非常缓慢，也有可能导致内存溢出，请不要直接调用
+    // this._decodeKeyBlock();
+
+    // @depreciated
+    // 因为mdx的索引存在排序问题，很多词条无法通过key-info检索到，原本计划将词条全部读取出来再重排序
+    // because the MDX's index has order issue, lots of words cannot be found by key-info
+    // so we have to read all words and resort them. However, it often causes memory overflow and
+    // very slow, so we have to avoid it.
+    // this._decodeKeyBlock();
+    // this._resortKeyList();
+    //
+    // 因为上述原因，resort选项也被弃用
+    // because of above reason, resort option also has been aborted
+    // if (!this.options.resort) {
+    //   throw new Error(
+    //     'js-mdict version above 5.0.0, must specify `options.resort = true` '
+    //   );
+    // }
 
     // -------------------------
     // dict record header section
     // --------------------------
     this._recordHeaderStartOffset = 0;
     this._recordHeaderEndOffset = 0;
-
-    this.keyList = [];
-
-    // depreciated part
-    // decodeKeyBlock method is very slow, avoid invoke dirctly
-    // this method will return the whole words list of the dictionaries file, this is very slow
-    // operation, and you should do this background, or concurrently.
-    // NOTE: this method is wrapped by method medict.RangeWords();
-    // this._decodeKeyBlock();
-    if (!this.options.resort) {
-      throw new Error(
-        'js-mdict version above 5.0.0, must specify `options.resort = true` '
-      );
-    }
-
-    // resort mdd files
-    if (this.options.debug) {
-      const memTrace = measure.measureMemFn();
-      memTrace('before resort decode');
-      measure.measureTimeFn(this, this._decodeKeyBlock)();
-      measure.measureTimeFn(this, this._resortKeyList)();
-      memTrace('after resort decode');
-    } else {
-      this._decodeKeyBlock();
-      this._resortKeyList();
-    }
-
     this.recordHeader = {
       recordBlocksNum: 0,
       entriesNum: 0,
       recordBlockInfoCompSize: 0,
       recordBlockCompSize: 0,
     };
+
     this._decodeRecordHeader();
+
+    if (this.options.debug) {
+      console.log('decodeRecordHeader finished', {
+        recordHeader: this.recordHeader,
+      });
+    }
 
     // -------------------------
     // dict record info section
     // --------------------------
-    this._recordInfoStartOffset = 0;
-    this._recordInfoEndOffset = 0;
+    this._recordBlockInfoStartOffset = 0;
+    this._recordBlockInfoEndOffset = 0;
     this.recordBlockInfoList = [];
     this._decodeRecordInfo();
+
+    if (this.options.debug) {
+      const tempRecordBlockInfoList = this.recordBlockInfoList.slice(0, 10);
+      console.debug('decodeRecordInfo finished', {
+        recordBlockInfoList_0_10: tempRecordBlockInfoList,
+      });
+    }
 
     // -------------------------
     // dict record block section
     // --------------------------
     this._recordBlockStartOffset = 0;
     this._recordBlockEndOffset = 0;
-    this.keyData = [];
+    this.recordBlockDataList = [];
+
     // decodeRecordBlock method is very slow, avoid invoke dirctly
+    // this method will return the whole words list of the dictionaries file, this is very slow
+    // operation, and you should do this background, or concurrently.
+    // decodeRecordBlock() 非常缓慢，不要直接调用
     // this._decodeRecordBlock();
 
     // ---------------------------------
@@ -162,7 +342,7 @@ class MDictBase {
   }
 
   /**
-   * STEP 1. read diction header
+   * STEP 1. read dictionary header
    * Get mdx header info (xml content to object)
    * [0:4], 4 bytes header length (header_byte_size), big-endian, 4 bytes, 16 bits
    * [4:header_byte_size + 4] header_bytes
@@ -176,11 +356,13 @@ class MDictBase {
     const header_size_buffer = this._readBuffer(0, 4);
     const headerByteSize = common.readNumber(
       header_size_buffer,
-      common.NUMFMT_UINT32
+      common.NUMFMT_UINT32 as NumFmt
     );
 
     // [4:header_byte_size + 4] header_bytes
-    const headerBuffer = readChunk.sync(this.fname, 4, headerByteSize);
+    const headerBuffer = Buffer.from(
+      readChunkSync(this.fname, 4, headerByteSize)
+    );
 
     // TODO: SKIP 4 bytes alder32 checksum
     // header_b_cksum should skip for now, because cannot get alder32 sum by js
@@ -242,10 +424,10 @@ class MDictBase {
     this._version = parseFloat(this.header.GeneratedByEngineVersion);
     if (this._version >= 2.0) {
       this._numWidth = 8;
-      this._numFmt = common.NUMFMT_UINT64;
+      this._numFmt = common.NUMFMT_UINT64 as NumFmt;
     } else {
       this._numWidth = 4;
-      this._numFmt = common.NUMFMT_UINT32;
+      this._numFmt = common.NUMFMT_UINT32 as NumFmt;
     }
     if (!this.header.Encoding || this.header.Encoding == '') {
       this._encoding = UTF8;
@@ -422,25 +604,24 @@ class MDictBase {
    * and decode the first key and last key infomation, etc.
    * @param {Buffer} keyBlockInfoBuff key block info buffer
    */
-  _decodeKeyBlockInfo(keyBlockInfoBuff) {
+  _decodeKeyBlockInfo(keyBlockInfoBuff: Buffer): KeyBlockInfo[] {
     const keyBlockNum = this.keyHeader.keyBlocksNum;
     const num_entries = this.keyHeader.entriesNum;
-    let kbInfoBuff;
+    let kbInfoBuff: Buffer;
     if (this._version >= 2.0) {
       // zlib compression
       assert(
-        keyBlockInfoBuff.slice(0, 4).toString('hex') === '02000000',
+        keyBlockInfoBuff.subarray(0, 4).toString('hex') === '02000000',
         'the compress type zlib should start with 0x02000000'
       );
-      let kbInfoCompBuff;
+      let kbInfoCompBuff: Buffer = keyBlockInfoBuff;
       if (this._encrypt === 2) {
         kbInfoCompBuff = common.mdxDecrypt(keyBlockInfoBuff);
-      } else if (this._encrypt === 0) {
-        kbInfoCompBuff = keyBlockInfoBuff;
       }
+      
       // For version 2.0, will compress by zlib, lzo just just for 1.0
       // key_block_info_compressed[0:8] => compress_type
-      kbInfoBuff = pako.inflate(kbInfoCompBuff.slice(8, kbInfoCompBuff.length));
+      kbInfoBuff = Buffer.from(zlib.inflateSync(kbInfoCompBuff.subarray(8)));
 
       // TODO: check the alder32 checksum
       // adler32 = unpack('>I', key_block_info_compressed[4:8])[0]
@@ -455,19 +636,19 @@ class MDictBase {
       kbInfoBuff = keyBlockInfoBuff;
     }
 
-    const key_block_info_list = [];
+    const key_block_info_list: KeyBlockInfo[] = [];
 
     // init tmp variables
     let countEntriesNum = 0;
-    let byteFmt = common.NUMFMT_UINT16;
+    let byteFmt = common.NUMFMT_UINT16 as NumFmt;
     let byteWidth = 2;
     let textTerm = 1;
     if (this._version >= 2.0) {
-      byteFmt = common.NUMFMT_UINT16;
+      byteFmt = common.NUMFMT_UINT16 as NumFmt;
       byteWidth = 2;
       textTerm = 1;
     } else {
-      byteFmt = common.NUMFMT_UINT8;
+      byteFmt = common.NUMFMT_UINT8 as NumFmt;
       byteWidth = 1;
       textTerm = 0;
     }
@@ -488,7 +669,7 @@ class MDictBase {
 
       const firstKeySize = common.readNumber(
         kbInfoBuff.slice(i, i + byteWidth),
-        byteFmt
+        byteFmt as NumFmt
       );
       i += byteWidth;
 
@@ -592,10 +773,20 @@ class MDictBase {
    * @param {string} phrase searching phrase
    * @param {function} stripfunc strip key string to compare
    */
-  _reduceWordKeyBlock(phrase, _s, compareFn) {
+  _reduceWordKeyBlock(
+    phrase: string,
+    _s?: (word: string) => string,
+    compareFn?: (a: string, b: string) => number
+  ): number {
+    if (phrase == '') {
+      return -1;
+    }
+    if (!compareFn) {
+      compareFn = (a: string, b: string) => a.localeCompare(b);
+    }
     if (!_s || _s == undefined) {
       // eslint-disable-next-line
-      _s = (word) => {
+      _s = (word: string) => {
         return word;
       };
     }
@@ -632,19 +823,35 @@ class MDictBase {
    * Note: this method runs very slow, please do not use this unless special target
    */
   _decodeKeyBlock() {
+    console.log('_decodeKeyBlock', 'before read buffer', {
+      keyBlockInfoEndOffset: this._keyBlockInfoEndOffset,
+      keyBlocksTotalSize: this.keyHeader.keyBlocksTotalSize,
+    });
+
     this._keyBlockStartOffset = this._keyBlockInfoEndOffset;
+
     const kbCompBuff = this._readBuffer(
       this._keyBlockStartOffset,
       this.keyHeader.keyBlocksTotalSize
     );
 
-    let key_list = [];
+    console.log('_decodeKeyBlock', 'read buffer sample', {
+      kbCompBuff: kbCompBuff.subarray(0, 100),
+    });
+    let key_list: KeyListItem[] = [];
     let kbStartOfset = 0;
+    console.log('_decodeKeyBlock', 'start decode key block', {
+      keyBlocksTotalSize: this.keyHeader.keyBlocksTotalSize,
+      kbCompBuff: kbCompBuff.subarray(0, 100),
+      keyBlockInfoListLength: this.keyBlockInfoList.length,
+    });
     // harvest keyblocks
     for (let idx = 0; idx < this.keyBlockInfoList.length; idx++) {
       const compSize = this.keyBlockInfoList[idx].keyBlockCompSize;
-      const decompressed_size = this.keyBlockInfoList[idx].keyBlockDecompSize;
+      const decompressedSize = this.keyBlockInfoList[idx].keyBlockDecompSize;
+
       const start = kbStartOfset;
+
       assert(
         start === this.keyBlockInfoList[idx].keyBlockCompAccumulator,
         'should be equal'
@@ -652,29 +859,47 @@ class MDictBase {
 
       const end = kbStartOfset + compSize;
       // 4 bytes : compression type
-      const kbCompType = new BufferList(kbCompBuff.slice(start, start + 4));
+      const kbCompType = Buffer.from(kbCompBuff.subarray(start, start + 4));
+      console.log('_decodeKeyBlock', 'decode key block', {
+        idx,
+        compSize,
+        decompressedSize,
+        start,
+        end,
+        kbCompType,
+        compressTypeHex: kbCompType.toString('hex'),
+      });
       // TODO 4 bytes adler32 checksum
       // # 4 bytes : adler checksum of decompressed key block
       // adler32 = unpack('>I', key_block_compressed[start + 4:start + 8])[0]
 
-      let key_block;
+      let key_block: Buffer;
       if (kbCompType.toString('hex') == '00000000') {
         key_block = kbCompBuff.slice(start + 8, end);
       } else if (kbCompType.toString('hex') == '01000000') {
-        // # decompress key block
-        const header = new ArrayBuffer([0xf0, decompressed_size]);
+        // # 解压缩键块
+        const header = new Uint8Array([0xf0, decompressedSize]);
+
+        const combinedKbCompressedBuff = common.appendBuffer(
+          header,
+          kbCompBuff.subarray(start + 8, end)
+        );
+        // console.log("============");
+        // console.log(combinedKbCompressedBuff);
+
+        // fs.appendFileSync("./data.txt", combinedKbCompressedBuff);
+        // console.log("============");
+
         const keyBlock = lzo1x.decompress(
-          common.appendBuffer(header, kbCompBuff.slice(start + 8, end)),
-          decompressed_size,
-          1308672
+          combinedKbCompressedBuff,
+          decompressedSize,
+          0 //  TODO ignore the init size
         );
-        key_block = bufferToArrayBuffer(keyBlock).slice(
-          keyBlock.byteOffset,
-          keyBlock.byteOffset + keyBlock.byteLength
-        );
+
+        key_block = Buffer.from(keyBlock);
       } else if (kbCompType.toString('hex') === '02000000') {
         // decompress key block
-        key_block = pako.inflate(kbCompBuff.slice(start + 8, end));
+        key_block =  Buffer.from(pako.inflate(kbCompBuff.subarray(start + 8, end)));
         // extract one single key block into a key list
         // notice that adler32 returns signed value
         // TODO compare with privious word
@@ -684,7 +909,8 @@ class MDictBase {
           `cannot determine the compress type: ${kbCompType.toString('hex')}`
         );
       }
-      const splitedKey = this._splitKeyBlock(new BufferList(key_block), idx);
+
+      const splitedKey = this._splitKeyBlock(Buffer.from(key_block), idx);
       key_list = key_list.concat(splitedKey);
       kbStartOfset += compSize;
     }
@@ -700,7 +926,7 @@ class MDictBase {
    * decode key block by key block id (from key info list)
    * @param {*} kbid key block id
    */
-  _decodeKeyBlockByKBID(kbid) {
+  _decodeKeyBlockByKBID(kbid: number): KeyListItem[] {
     this._keyBlockStartOffset = this._keyBlockInfoEndOffset;
     const compSize = this.keyBlockInfoList[kbid].keyBlockCompSize;
     const decompSize = this.keyBlockInfoList[kbid].keyBlockDecompSize;
@@ -710,29 +936,29 @@ class MDictBase {
     const kbCompBuff = this._readBuffer(startOffset, compSize);
     const start = 0;
     const end = compSize;
-    const kbCompType = new BufferList(kbCompBuff.slice(start, start + 4));
+    const kbCompType = Buffer.from(kbCompBuff.subarray(start, start + 4));
     // TODO 4 bytes adler32 checksum
     // # 4 bytes : adler checksum of decompressed key block
     // adler32 = unpack('>I', key_block_compressed[start + 4:start + 8])[0]
 
-    let key_block;
+    let key_block: Buffer;
     if (kbCompType.toString('hex') == '00000000') {
-      key_block = kbCompBuff.slice(start + 8, end);
+      key_block = kbCompBuff.subarray(start + 8, end);
     } else if (kbCompType.toString('hex') == '01000000') {
       // # decompress key block
-      const header = new ArrayBuffer([0xf0, decompSize]);
       const keyBlock = lzo1x.decompress(
-        common.appendBuffer(header, kbCompBuff.slice(start + 8, end)),
+        kbCompBuff.subarray(start + 8, end),
         decompSize,
         1308672
       );
-      key_block = bufferToArrayBuffer(keyBlock).slice(
+
+      key_block = Buffer.from(keyBlock).subarray(
         keyBlock.byteOffset,
         keyBlock.byteOffset + keyBlock.byteLength
       );
     } else if (kbCompType.toString('hex') === '02000000') {
       // decompress key block
-      key_block = pako.inflate(kbCompBuff.slice(start + 8, end));
+      key_block =  Buffer.from(pako.inflate(kbCompBuff.slice(start + 8, end)));
       // extract one single key block into a key list
       // notice that adler32 returns signed value
       // TODO compare with privious word
@@ -742,7 +968,7 @@ class MDictBase {
         `cannot determine the compress type: ${kbCompType.toString('hex')}`
       );
     }
-    const splitedKey = this._splitKeyBlock(new BufferList(key_block), kbid);
+    const splitedKey = this._splitKeyBlock(Buffer.from(key_block), kbid);
     return splitedKey;
   }
 
@@ -751,17 +977,17 @@ class MDictBase {
    * split key from key block buffer
    * @param {Buffer} keyBlock key block buffer
    */
-  _splitKeyBlock(keyBlock, keyBlockIdx) {
-    let delimiter;
-    let width;
+  _splitKeyBlock(keyBlock: Buffer, keyBlockIdx: number): KeyListItem[] {
+    // let delimiter: string = '00';
+    let width: number;
     if (this._encoding == 'UTF-16' || this.ext == 'mdd') {
-      delimiter = '0000';
+      // delimiter = '0000';
       width = 2;
     } else {
-      delimiter = '00';
+      // delimiter = '00';
       width = 1;
     }
-    const keyList = [];
+    const keyList: KeyListItem[] = [];
     let keyStartIndex = 0;
     let keyEndIndex = 0;
 
@@ -769,7 +995,7 @@ class MDictBase {
       // # the corresponding record's offset in record block
       // 0.2656s
       const recordStartOffset = common.readNumber(
-        keyBlock.slice(keyStartIndex, keyStartIndex + this._numWidth),
+        keyBlock.subarray(keyStartIndex, keyStartIndex + this._numWidth),
         this._numFmt
       );
 
@@ -783,13 +1009,13 @@ class MDictBase {
       while (i < keyBlock.length) {
         // delimiter = '0'
         if (
-          (width === 1 && keyBlock.get(i) == 0) ||
+          (width === 1 && keyBlock[i] == 0) ||
           // delimiter = '00'
-          (width === 2 && keyBlock.get(i) == 0 && keyBlock.get(i + 1) == 0)
+          (width === 2 && keyBlock[i] == 0 && keyBlock[i + 1] == 0)
         ) {
           //// the method below was very slow, depreate
           // if (
-          //   new BufferList(keyBlock.slice(i, i + width)).toString('hex') ==
+          //  Buffer.from(keyBlock.slice(i, i + width)).toString('hex') ==
           //   delimiter
           // ) {
           keyEndIndex = i;
@@ -798,7 +1024,7 @@ class MDictBase {
         i += width;
       }
       const keyText = this._decoder.decode(
-        keyBlock.slice(keyStartIndex + this._numWidth, keyEndIndex)
+        keyBlock.subarray(keyStartIndex + this._numWidth, keyEndIndex)
       );
       keyStartIndex = keyEndIndex + width;
       keyList.push({ recordStartOffset, keyText, keyBlockIdx });
@@ -815,33 +1041,38 @@ class MDictBase {
    * [16:24/8:12] - record block info size
    * [24:32/12:16] - record block size
    */
-  _decodeRecordHeader() {
+  _decodeRecordHeader(): void {
     this._recordHeaderStartOffset =
       this._keyBlockInfoEndOffset + this.keyHeader.keyBlocksTotalSize;
     const rhlen = this._version >= 2.0 ? 4 * 8 : 4 * 4;
     this._recordHeaderEndOffset = this._recordHeaderStartOffset + rhlen;
     const rhBuff = this._readBuffer(this._recordHeaderStartOffset, rhlen);
+
     let ofset = 0;
     const recordBlocksNum = common.readNumber(
       rhBuff.slice(ofset, ofset + this._numWidth),
       this._numFmt
     );
+
     ofset += this._numWidth;
     const entriesNum = common.readNumber(
       rhBuff.slice(ofset, ofset + this._numWidth),
       this._numFmt
     );
     assert(entriesNum === this.keyHeader.entriesNum);
+
     ofset += this._numWidth;
     const recordBlockInfoCompSize = common.readNumber(
       rhBuff.slice(ofset, ofset + this._numWidth),
       this._numFmt
     );
+
     ofset += this._numWidth;
     const recordBlockCompSize = common.readNumber(
       rhBuff.slice(ofset, ofset + this._numWidth),
       this._numFmt
     );
+
     this.recordHeader = {
       recordBlocksNum,
       entriesNum,
@@ -860,10 +1091,10 @@ class MDictBase {
    * decomAccu
    * }]
    */
-  _decodeRecordInfo() {
-    this._recordInfoStartOffset = this._recordHeaderEndOffset;
+  _decodeRecordInfo(): void {
+    this._recordBlockInfoStartOffset = this._recordHeaderEndOffset;
     const riBuff = this._readBuffer(
-      this._recordInfoStartOffset,
+      this._recordBlockInfoStartOffset,
       this.recordHeader.recordBlockInfoCompSize
     );
     /**
@@ -876,7 +1107,7 @@ class MDictBase {
      * }]
      * Note: every record block will contrains a lot of entries
      */
-    const recordBlockInfoList = [];
+    const recordBlockInfoList: RecordBlockInfo[] = [];
     let ofset = 0;
     let compAccu = 0;
     let decompAccu = 0;
@@ -904,10 +1135,11 @@ class MDictBase {
     assert(ofset === this.recordHeader.recordBlockInfoCompSize);
     assert(compAccu === this.recordHeader.recordBlockCompSize);
     this.recordBlockInfoList = recordBlockInfoList;
-    this._recordInfoEndOffset =
-      this._recordInfoStartOffset + this.recordHeader.recordBlockInfoCompSize;
+    this._recordBlockInfoEndOffset =
+      this._recordBlockInfoStartOffset +
+      this.recordHeader.recordBlockInfoCompSize;
     // avoid user not invoke the _decodeRecordBlock method
-    this._recordBlockStartOffset = this._recordInfoEndOffset;
+    this._recordBlockStartOffset = this._recordBlockInfoEndOffset;
   }
 
   /**
@@ -915,9 +1147,9 @@ class MDictBase {
    * decode all records block,
    * this is a slowly method, do not use!
    */
-  _decodeRecordBlock() {
-    this._recordBlockStartOffset = this._recordInfoEndOffset;
-    const keyData = [];
+  _decodeRecordBlock(): void {
+    this._recordBlockStartOffset = this._recordBlockInfoEndOffset;
+    const keyData: any[] = [];
 
     /**
      * start reading the record block
@@ -935,10 +1167,10 @@ class MDictBase {
       recordOffset += compSize;
 
       // 4 bytes: compression type
-      const rbCompType = new BufferList(rbCompBuff.slice(0, 4));
+      const rbCompType = Buffer.from(rbCompBuff.slice(0, 4));
 
       // record_block stores the final record data
-      let recordBlock;
+      let recordBlock: Buffer = Buffer.from([]);
 
       // TODO: igore adler32 offset
       // Note: here ignore the checksum part
@@ -950,7 +1182,7 @@ class MDictBase {
         // --------------
         // decrypt
         // --------------
-        let blockBufDecrypted = null;
+        let blockBufDecrypted: Buffer | null = null;
         // if encrypt type == 1, the record block was encrypted
         if (
           this._encrypt ===
@@ -969,24 +1201,26 @@ class MDictBase {
         if (rbCompType.toString('hex') === '01000000') {
           comp_type = 'lzo';
           // the header was need by lzo library, should append before real compressed data
-          const header = new ArrayBuffer([0xf0, decompSize]);
+          const header = Buffer.from([0xf0, decompSize]);
           // Note: if use lzo, here will LZO_E_OUTPUT_RUNOVER, so ,use mini lzo js
-          recordBlock = lzo1x.decompress(
-            common.appendBuffer(header, blockBufDecrypted),
-            decompSize,
-            1308672
+          recordBlock = Buffer.from(
+            lzo1x.decompress(
+              common.appendBuffer(header, blockBufDecrypted),
+              decompSize,
+              1308672
+            )
           );
-          recordBlock = bufferToArrayBuffer(recordBlock).slice(
+          recordBlock = Buffer.from(recordBlock).slice(
             recordBlock.byteOffset,
             recordBlock.byteOffset + recordBlock.byteLength
           );
         } else if (rbCompType.toString('hex') === '02000000') {
           comp_type = 'zlib';
           // zlib decompress
-          recordBlock = pako.inflate(blockBufDecrypted);
+          recordBlock =  Buffer.from(pako.inflate(blockBufDecrypted));
         }
       }
-      recordBlock = new BufferList(recordBlock);
+      recordBlock = Buffer.from(recordBlock);
 
       // notice that adler32 return signed value
       // TODO: ignore the checksum
@@ -1012,7 +1246,7 @@ class MDictBase {
           break;
         }
         // # record end index
-        let recordEnd;
+        let recordEnd: number;
         if (i < this.keyList.length - 1) {
           recordEnd = this.keyList[i + 1].recordStartOffset;
         } else {
@@ -1045,7 +1279,7 @@ class MDictBase {
 
     assert(sizeCounter === this.recordHeader.recordBlockCompSize);
 
-    this.keyData = keyData;
+    this.recordBlockDataList = keyData;
     this._recordBlockEndOffset = this._recordBlockStartOffset + sizeCounter;
   }
 
@@ -1053,7 +1287,7 @@ class MDictBase {
    * find record which record start locate
    * @param {number} recordStart record start offset
    */
-  _reduce_record_block(recordStart) {
+  _reduceRecordBlockInfo(recordStart: number): number {
     let left = 0;
     let right = this.recordBlockInfoList.length - 1;
     let mid = 0;
@@ -1075,9 +1309,14 @@ class MDictBase {
    * @param {number} start this word record start offset
    * @param {number} nextStart next word record start offset
    */
-  _decode_record_block_by_rb_id(rbid, keyText, start, nextStart) {
+  _decodeRecordBlockDataByRecordBlockInfoId(
+    rbid: number,
+    keyText: string,
+    start: number,
+    nextStart: number
+  ): { keyText: string; definition: string } {
     // decode record block by record block id
-    this._recordBlockStartOffset = this._recordInfoEndOffset;
+    this._recordBlockStartOffset = this._recordBlockInfoEndOffset;
     const compSize = this.recordBlockInfoList[rbid].compSize;
     const decompSize = this.recordBlockInfoList[rbid].decompSize;
     const compAccumulator = this.recordBlockInfoList[rbid].compAccumulator;
@@ -1086,22 +1325,22 @@ class MDictBase {
     const rbCompBuff = this._readBuffer(startOffset, compSize);
 
     // 4 bytes: compression type
-    const rbCompType = new BufferList(rbCompBuff.slice(0, 4));
+    const rbCompType = Buffer.from(rbCompBuff.subarray(0, 4));
 
     // record_block stores the final record data
-    let recordBlock;
+    let recordBlock: Buffer = Buffer.from([]);
 
     // TODO: igore adler32 offset
     // Note: here ignore the checksum part
     // bytes: adler32 checksum of decompressed record block
     // adler32 = unpack('>I', record_block_compressed[4:8])[0]
     if (rbCompType.toString('hex') === '00000000') {
-      recordBlock = rbCompBuff.slice(8, rbCompBuff.length);
+      recordBlock = rbCompBuff.subarray(8, rbCompBuff.length);
     } else {
       // --------------
       // decrypt
       // --------------
-      let blockBufDecrypted = null;
+      let blockBufDecrypted: Buffer | null = null;
       // if encrypt type == 1, the record block was encrypted
       if (
         this._encrypt === 1 /* || (this.ext == "mdd" && this._encrypt === 2 ) */
@@ -1111,30 +1350,36 @@ class MDictBase {
         // passkey.set([0x95, 0x36, 0x00, 0x00], 4); // key part 2: fixed data
         blockBufDecrypted = common.mdxDecrypt(rbCompBuff);
       } else {
-        blockBufDecrypted = rbCompBuff.slice(8, rbCompBuff.length);
+        blockBufDecrypted = rbCompBuff.subarray(8, rbCompBuff.length);
       }
       // --------------
       // decompress
       // --------------
       if (rbCompType.toString('hex') === '01000000') {
         // the header was need by lzo library, should append before real compressed data
-        const header = new ArrayBuffer([0xf0, decompSize]);
-        // Note: if use lzo, here will LZO_E_OUTPUT_RUNOVER, so ,use mini lzo js
-        recordBlock = lzo1x.decompress(
-          common.appendBuffer(header, blockBufDecrypted),
-          decompSize,
-          1308672
-        );
-        recordBlock = bufferToArrayBuffer(recordBlock).slice(
+
+        // const header = Buffer.from([0xf0, decompSize]);
+        // // Note: if use lzo, here will LZO_E_OUTPUT_RUNOVER, so ,use mini lzo js
+        // recordBlock = Buffer.from(
+        //   lzo1x.decompress(
+        //     common.appendBuffer(header, blockBufDecrypted),
+        //     decompSize,
+        //     1308672
+        //   )
+        // );
+
+        recordBlock = lzo1x.decompress(blockBufDecrypted, decompSize, 1308672);
+
+        recordBlock = Buffer.from(recordBlock).subarray(
           recordBlock.byteOffset,
           recordBlock.byteOffset + recordBlock.byteLength
         );
       } else if (rbCompType.toString('hex') === '02000000') {
         // zlib decompress
-        recordBlock = pako.inflate(blockBufDecrypted);
+        recordBlock =  Buffer.from(pako.inflate(blockBufDecrypted));
       }
     }
-    recordBlock = new BufferList(recordBlock);
+    recordBlock = Buffer.from(recordBlock);
 
     // notice that adler32 return signed value
     // TODO: ignore the checksum
@@ -1143,19 +1388,19 @@ class MDictBase {
 
     const recordStart = start - decompAccumulator;
     const recordEnd = nextStart - decompAccumulator;
-    const data = recordBlock.slice(recordStart, recordEnd);
+    const data = recordBlock.subarray(recordStart, recordEnd);
     if (this.ext === 'mdd') {
       return { keyText, definition: BASE64ENCODER(data) };
     }
     return { keyText, definition: this._decoder.decode(data) };
   }
 
-  _readBuffer(start, length) {
-    return readChunk.sync(this.fname, start, length);
+  _readBuffer(start: number, length: number): Buffer {
+    return Buffer.from(readChunkSync(this.fname, start, length));
   }
 
   // store key to wordBuffer
-  _resortKeyList() {
+  _resortKeyList(): void {
     // 排序之前记录下，每个单词的结束位置，因为排序之后顺序就乱了，buffer 里就不能再根据下一个单词判断了
     this.keyList.map((v, i) => {
       v.original_idx = i;
@@ -1167,41 +1412,209 @@ class MDictBase {
     this.keyListRemap = {};
 
     if (this._isKeyCaseSensitive()) {
-      this.keyList.sort(common.caseUnsensitiveCompare);
+      this.keyList.sort((a, b) =>
+        common.caseUnsensitiveCompare(a.keyText, b.keyText)
+      );
     } else {
-      this.keyList.sort(common.caseSensitiveCompare);
+      this.keyList.sort((a, b) =>
+        common.caseSensitiveCompare(a.keyText, b.keyText)
+      );
     }
 
     // build index remap
-    this.keyList.map((v, i) => {
-      this.keyListRemap[v.original_idx] = i;
+    this.keyList.forEach((v, i) => {
+      this.keyListRemap[v.original_idx!] = i;
     });
   }
 
-  _strip_key_or_ingore_case() {
-    return function (key) {
-      // this strip/case sensistive part will increase time cost about 100% (20ms->38ms)
-      if (this._isStripKey()) {
-        key = key.replace(common.REGEXP_STRIPKEY[this.ext], '$1');
-      }
-      if (!this._isKeyCaseSensitive()) {
-        key = key.toLowerCase();
-      }
-      if (this.ext == 'mdd') {
-        key = key.replace(/\\/g, '/');
-      }
-      return key.trim();
-    }.bind(this);
+  _stripKeyOrIngoreCase(key: string): string {
+    // this strip/case sensistive part will increase time cost about 100% (20ms->38ms)
+    if (this._isStripKey()) {
+      key = key.replace(common.REGEXP_STRIPKEY[this.ext], '$1');
+    }
+    if (!this._isKeyCaseSensitive()) {
+      key = key.toLowerCase();
+    }
+    if (this.ext == 'mdd') {
+      key = key.replace(/\\/g, '/');
+    }
+    return key.trim();
   }
 
-  _isKeyCaseSensitive() {
+  _isKeyCaseSensitive(): boolean {
     return (
       this.options.isCaseSensitive || common.isTrue(this.header.isCaseSensitive)
     );
   }
 
-  _isStripKey() {
+  _isStripKey(): boolean {
     return this.options.isStripKey || common.isTrue(this.header.StripKey);
+  }
+
+  _lookupKeyBlockId(word: string) {
+    const lookupInternal = (compareFn: (a: string, b: string) => number) => {
+      const sfunc = this._stripKeyOrIngoreCase.bind(this);
+      const kbid = this._reduceWordKeyBlock(word, sfunc, compareFn);
+      // not found
+      if (kbid < 0) {
+        return undefined;
+      }
+      const list = this._decodeKeyBlockByKBID(kbid);
+      const i = this._binarySearh(list, word, sfunc, compareFn);
+      if (i === undefined) {
+        return undefined;
+      }
+      return { idx: i, list };
+    };
+
+    let record;
+    if (this._isKeyCaseSensitive()) {
+      record = lookupInternal(common.normalUpperCaseWordCompare);
+    } else {
+      record = lookupInternal(common.normalUpperCaseWordCompare);
+      if (record === undefined) {
+        record = lookupInternal(common.wordCompare);
+      }
+    }
+    return record;
+  }
+
+  _binarySearh(
+    list: KeyListItem[],
+    word: string,
+    _s: (s: string) => string,
+    compareFn: (a: string, b: string) => number
+  ) {
+    let left = 0;
+    let right = list.length - 1;
+    let mid = 0;
+    while (left <= right) {
+      mid = left + ((right - left) >> 1);
+      // if case sensitive, the uppercase word is smaller than lowercase word
+      // for example: `Holanda` is smaller than `abacaxi`
+      // so when comparing with the words, we should use the dictionary order,
+      // however, if we change the word to lowercase, the binary search algorithm will be confused
+      // so, we use the enhanced compare function `common.wordCompare`
+      const compareResult = compareFn(_s(word), _s(list[mid].keyText));
+      // console.log(`@#@# wordCompare ${_s(word)} ${_s(list[mid].keyText)} ${compareResult} l: ${left} r: ${right} mid: ${mid} ${list[mid].keyText}`)
+      if (compareResult > 0) {
+        left = mid + 1;
+      } else if (compareResult == 0) {
+        return mid;
+      } else {
+        right = mid - 1;
+      }
+    }
+    return mid;
+  }
+
+  _search_key_record(word: string): KeyRecord | undefined {
+    word = this._stripKeyOrIngoreCase(word);
+    // use js internal find
+    const keyBlockInfo = this.keyBlockInfoList.find((item, index) => {
+      if (item.firstKey <= word && item.lastKey >= word) {
+        return true;
+      }
+    });
+    if (!keyBlockInfo) {
+      return undefined;
+    }
+    let propRecordBlockInfo: RecordBlockInfo | undefined = undefined;
+    let propIdx = 0;
+    for (let i = 0; i < this.recordBlockInfoList.length; i++) {
+      const item = this.recordBlockInfoList[i];
+      if (item.decompAccumulator >= keyBlockInfo.keyBlockDecompAccumulator) {
+        propRecordBlockInfo = item;
+        propIdx = i;
+        break;
+      }
+    }
+
+    if (propRecordBlockInfo == undefined) {
+      return undefined;
+    }
+
+    let locatedRecordBlockInfo = propRecordBlockInfo;
+    let locatedIdx = propIdx;
+    // 处理正常情况
+    if (propIdx > 0 && propIdx < this.recordBlockDataList.length - 1) {
+      locatedRecordBlockInfo = this.recordBlockDataList[propIdx - 1];
+      locatedIdx = propIdx - 1;
+    } else if (propIdx == 0) {
+      const firstBlockInfo = this.recordBlockDataList[0];
+
+      const lastBlockInfo =
+        this.recordBlockDataList[this.recordBlockDataList.length - 1];
+
+      // 处理第一个block
+      if (firstBlockInfo.firstKey <= word && lastBlockInfo.lastKey >= word) {
+        locatedRecordBlockInfo = firstBlockInfo;
+        locatedIdx = 0;
+      }
+      // 处理最后一个block
+      if (lastBlockInfo.firstKey <= word && lastBlockInfo.lastKey >= word) {
+        locatedRecordBlockInfo = lastBlockInfo;
+        locatedIdx = this.recordBlockDataList.length - 1;
+      }
+    } else {
+      // 其他情况不可能出现
+    }
+
+    const result: KeyListItem = {
+      keyText: word,
+      recordStartOffset: locatedRecordBlockInfo.decompAccumulator,
+      nextRecordStartOffset:
+        locatedRecordBlockInfo.decompAccumulator +
+        locatedRecordBlockInfo.decompSize,
+      original_idx: locatedIdx,
+
+      keyBlockIdx: locatedIdx,
+    };
+    return result;
+  }
+
+  _locate_prefix(word: string): number {
+    const end = this.keyList.length;
+    word = this._stripKeyOrIngoreCase(word);
+    for (let i = 0; i < end; i++) {
+      const keyText = this._stripKeyOrIngoreCase(this.keyList[i].keyText);
+      if (keyText.startsWith(word)) {
+        return i;
+      }
+    }
+    return -1;
+  }
+
+  _locate_prefix_list(
+    phrase: string,
+    max_len = 100,
+    max_missed = 100
+  ): KeyRecord[] {
+    const record = this._locate_prefix(phrase);
+    if (record == -1) {
+      return [];
+    }
+    const fn = this._stripKeyOrIngoreCase;
+
+    const list: KeyRecord[] = [];
+    let count = 0;
+    let missed = 0;
+    for (let i = record; i < this.keyList.length; i++) {
+      if (this.keyList[i].keyText.startsWith(fn(phrase))) {
+        list.push(this.keyList[i]);
+        count++;
+      } else {
+        missed++;
+      }
+      if (count > max_len) {
+        break;
+      }
+      if (missed > max_missed) {
+        break;
+      }
+    }
+
+    return list;
   }
 
   /**
